Update on app
1. Worked on the business and talent job with event controllers to properly assign the values to their respective units.

Due to the BE (Elijah) code, where he returns just strings of ID for the job applicant and savers and in other cases returns the full applicant's details, having a reusable component is now something stressful. I will have to rebuild or even duplicate, as the 'JobCard' and the 'SavedJobcard' is really confusing me at this state.